[MQTT5] Pull out property code
